fixes of hashCode and tests

+++ libraries/stdlib/src/kotlin/ranges/Progressions.kt
@@ -179,7 +179,7 @@ public open class LongProgression
         first == other.first && last == other.last && step == other.step)
 
     override fun hashCode(): Int =
-        if (isEmpty()) -1 else (31 * (31 * first.hashCode() + last.hashCode()) + step.hashCode()).toInt()
+        if (isEmpty()) -1 else (31 * (31 * first.hashCode() + last.hashCode()) + step.hashCode())
 
     override fun toString(): String = if (step > 0) "$first..$last step $step" else "$first downTo $last step ${-step}"
